requirements.txt
# Core dependencies
requests>=2.25.0
pycryptodome>=3.15.0
pyperclip>=1.8.0

# Optional: SEED encryption (Korean standard)
# Uncomment if you need SEED crypto with better performance
# kisa-seed>=1.0.0
